CSP-1088 Moved domain classes with in domain folder
