+++ src/AdminAan/SFA.DAS.AdminAan.UnitTests/Application/CalendarEvents/Commands/Create/PostEventCommandHandlerTests.cs
@@ -2,6 +2,7 @@
 using FluentAssertions;
 using Moq;
 using SFA.DAS.AdminAan.Application.CalendarEvents.Commands.Create;
+using SFA.DAS.AdminAan.Domain;
 using SFA.DAS.AdminAan.Infrastructure;
 using SFA.DAS.Testing.AutoFixture;
