+++ src/AdminAan/SFA.DAS.AdminAan/Application/CalendarEvents/Commands/Create/GuestSpeakers.cs
@@ -1,4 +1,4 @@
-﻿using SFA.DAS.AdminAan.Infrastructure;
+﻿using SFA.DAS.AdminAan.Domain;
 
 namespace SFA.DAS.AdminAan.Application.CalendarEvents.Commands.Create;
 public record PutEventGuestsModel(List<GuestSpeaker> Guests);
